사이드바 연동 뜯어고치기 전 커밋
